pyproject.toml
[tool.poetry]
name = "nonebot-plugin-onebot-monitor"
version = "0.1.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
license = "MIT"
packages = [
    { include = "nonebot_plugin_onebot_monitor", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-rc.1"
nonebot-adapter-onebot = "^2.1.3"
nonebot-plugin-sqlalchemy = "^0.1.0"
aiosqlite = "^0.17.0"

[tool.poetry.group.dev.dependencies]
sqlalchemy2-stubs = "^0.0.2a29"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
